+++ backend/src/main/java/com/optimize/elykia/core/service/user/UserManagement.java
@@ -2,8 +2,10 @@ package com.optimize.elykia.core.service.user;
 
 import com.optimize.common.securities.models.User;
 import com.optimize.common.securities.security.services.UserService;
+import com.optimize.elykia.core.config.CacheNames;
 import com.optimize.elykia.core.util.UserProfilConstant;
 import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -13,6 +15,7 @@ import java.util.*;
 public class UserManagement {
     private final UserService userService;
 
+    @Cacheable(cacheNames = CacheNames.PROMOTERS)
     public List<User> getPromoters() {
         return userService.getByUserProfil(UserProfilConstant.PROMOTER);
     }